+++ components/brave_rewards/resources/page/brave_rewards_page.tsx
@@ -20,7 +20,8 @@ import * as rewardsActions from './actions/rewards_actions'
 import { PlatformContext } from './lib/platform_context'
 import { createReducer } from './reducers'
 import { getCurrentBalanceReport } from './reducers/utils'
-import { App } from './components/app'
+// import { App } from './components/app'
+import { PdfRenderer } from './components/pdf_renderer'
 import * as mojom from '../shared/lib/mojom'
 
 import * as Rewards from './lib/types'
@@ -51,7 +52,7 @@ function initialize () {
         <LocaleContext.Provider value={createLocaleContextForWebUI()}>
           <PlatformContext.Provider value={platformInfo}>
             <WithThemeVariables>
-              <App />
+              <PdfRenderer />
               <></>
             </WithThemeVariables>
           </PlatformContext.Provider>
