+++ components/brave_rewards/resources/page/components/pdf_renderer.tsx
