+++ components/brave_rewards/resources/page/components/pdfjs-dist-worker.js
